fix1

+++ DB2ProjectEJB/ejbModule/it/polimi/db2/project/ejb/entities/AlertEntity.java
@@ -19,7 +19,8 @@ import javax.persistence.Table;
 @Entity
 @Table(name = "alert")
 @NamedQueries({
-        @NamedQuery(name = "AlertEntity.findUser", query = "SELECT u FROM UserEntity u WHERE u.id = :user_alert"),         
+        @NamedQuery(name = "AlertEntity.findUser", query = "SELECT u FROM UserEntity u WHERE u.id = :user_alert"), // TODO controllare sia corretta
+        
 })
 public class AlertEntity {
 
@@ -29,7 +30,7 @@ public class AlertEntity {
     private int id;
  
  	@Column(name = "Amount", nullable = false, length = 45)
-    private float amount;
+    private int amount;
 
     @Column(name = "Timestamp", nullable = false, length = 45) // put by default from the db when the entry is created in the table
     private Timestamp timestamp;
@@ -39,7 +40,7 @@ public class AlertEntity {
     private UserEntity user_alert;
 
     
-	    public AlertEntity(float amount, Timestamp timestamp, UserEntity user_alert) {
+	    public AlertEntity(int amount, Timestamp timestamp, UserEntity user_alert) {
 	    	this.amount = amount;
 	        this.timestamp = timestamp;
 	        this.user_alert = user_alert;	      
@@ -56,7 +57,7 @@ public class AlertEntity {
 		this.id = id;
 	}
 
-	public float getAmount() {
+	public int getAmount() {
 		return amount;
 	}
